Crate rustc_driver

The Rust compiler.

Note

This API is completely unstable and subject to change.

Re-exports

pub extern crate rustc_plugin_impl as plugin;

Modules

args

pretty

The various pretty-printing routines.

session_diagnostics 🔒

signal_handler 🔒

Structs

RunCompiler

This is the primary entry point for rustc.

TimePassesCallbacks

Enums

Compilation

Whether to stop or continue compilation.

Constants

BUG_REPORT_URL 🔒

EXIT_FAILURE

Exit status code used for compilation failures and invalid flags.

EXIT_SUCCESS

Exit status code used for successful compilation and help output.

ICE_REPORT_COMPILER_FLAGS 🔒

ICE_REPORT_COMPILER_FLAGS_EXCLUDE 🔒

ICE_REPORT_COMPILER_FLAGS_STRIP_VALUE 🔒

Statics

DEFAULT_HOOK 🔒

Traits

Callbacks

Functions

abort_on_err

catch_fatal_errors

Runs a closure and catches unwinds triggered by fatal errors.

catch_with_exit_code

Variant of catch_fatal_errors for the interface::Result return type that also computes the exit code.

describe_codegen_flags 🔒

describe_debug_flags 🔒

describe_lints

Write to stdout lint command options, together with a list of all available lints

diagnostics_registry

extra_compiler_flags 🔒

Gets a list of extra command-line flags provided by the user, as strings.

handle_explain 🔒

handle_options

Process command line options. Emits messages as appropriate. If compilation should continue, returns a getopts::Matches object parsed from args, otherwise returns None.

init_env_logger

This allows tools to enable rust logging without having to magically match rustc’s tracing crate version. In contrast to init_rustc_env_logger it allows you to choose an env var other than RUSTC_LOG.

init_rustc_env_logger

This allows tools to enable rust logging without having to magically match rustc’s tracing crate version.

install_ice_hook

Installs a panic hook that will print the ICE message on unexpected panics.

list_metadata

main

make_input 🔒

make_output 🔒

parse_crate_attrs 🔒

print_crate_info 🔒

print_flag_list

print_wall_help 🔒

report_ice

Prints the ICE message, including query stack, but without backtrace.

run_compiler 🔒

set_sigpipe_handler

show_content_with_pager 🔒

try_process_rlink

usage 🔒

version

Prints version information