Module rustc_driver::diagnostic

Unstable (rustc_private)

Structs

EmitterWriter 	 [Unstable]

ExplicitBug 	 [Unstable]
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError 	 [Unstable]
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler 	 [Unstable]
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

SpanHandler 	 [Unstable]
A span-handler is like a handler but also accepts span information for source-location reporting.

Enums

ColorConfig 	 [Unstable]

Level 	 [Unstable]

RenderSpan 	 [Unstable]

Traits

Emitter 	 [Unstable]

Functions

expect 	 [Unstable]