Struct rustc_driver::diagnostic::FatalErrorUnstable

pub struct FatalError;

This structure is used to signify that a task has panicked with a fatal error from the diagnostics. You can use this with the Any trait to figure out how a rustc task died (if so desired).