Function rustc_driver::driver::phase_2_configure_and_expand

pub fn phase_2_configure_and_expand<F>(
    sess: &Session, 
    cstore: &CStore, 
    krate: Crate, 
    registry: Option<Registry>, 
    crate_name: &str, 
    addl_plugins: Option<Vec<String>>, 
    make_glob_map: MakeGlobMap, 
    after_expand: F
) -> Result<ExpansionResult, CompileIncomplete> where
    F: FnOnce(&Crate) -> CompileResult, 

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Run the "early phases" of the compiler: initial cfg processing, loading compiler plugins (including those from addl_plugins), syntax expansion, secondary cfg expansion, synthesis of a test harness if one is to be provided, injection of a dependency on the standard library and prelude, and name resolution.

Returns None if we're aborting after handling -W help.