Function rustc_driver::driver::phase_3_run_analysis_passes

pub fn phase_3_run_analysis_passes<'tcx, F, R>(sess: &'tcx Session, ast_map: Map<'tcx>, arenas: &'tcx CtxtArenas<'tcx>, name: &str, make_glob_map: MakeGlobMap, f: F) -> R where F: for<'a> FnOnce(&'a ctxt<'tcx>, MirMap<'tcx>, CrateAnalysis) -> R

Unstable (rustc_private)

Run the resolution, typechecking, region checking and other miscellaneous analysis passes on the crate. Return various structures carrying the results of the analysis.