Function rustc_driver::driver::phase_6_link_output

pub fn phase_6_link_output(sess: &Session, trans: &CrateTranslation, outputs: &OutputFilenames)

Unstable

Run the linker on any artifacts that resulted from the LLVM run. This should produce either a finished executable or library.