Struct rustc_driver::driver::CompileState

pub struct CompileState<'a, 'b, 'ast: 'a, 'tcx: 'b> where 'ast: 'tcx {
    pub input: &'a Input,
    pub session: &'ast Session,
    pub krate: Option<Crate>,
    pub cstore: Option<&'a CStore>,
    pub crate_name: Option<&'a str>,
    pub output_filenames: Option<&'a OutputFilenames>,
    pub out_dir: Option<&'a Path>,
    pub out_file: Option<&'a Path>,
    pub arenas: Option<&'ast CtxtArenas<'ast>>,
    pub expanded_crate: Option<&'a Crate>,
    pub hir_crate: Option<&'a Crate>,
    pub ast_map: Option<&'a Map<'ast>>,
    pub resolutions: Option<&'a Resolutions>,
    pub mir_map: Option<&'b MirMap<'tcx>>,
    pub analysis: Option<&'a CrateAnalysis<'a>>,
    pub tcx: Option<TyCtxt<'b, 'tcx, 'tcx>>,
    pub trans: Option<&'a CrateTranslation>,
}

Unstable (rustc_private)

State that is passed to a callback. What state is available depends on when during compilation the callback is made. See the various constructor methods (state_*) in the impl to see which data is provided for any given entry point.

Fields

input
Unstable (rustc_private)

session
Unstable (rustc_private)

krate
Unstable (rustc_private)

cstore
Unstable (rustc_private)

crate_name
Unstable (rustc_private)

output_filenames
Unstable (rustc_private)

out_dir
Unstable (rustc_private)

out_file
Unstable (rustc_private)

arenas
Unstable (rustc_private)

expanded_crate
Unstable (rustc_private)

hir_crate
Unstable (rustc_private)

ast_map
Unstable (rustc_private)

resolutions
Unstable (rustc_private)

mir_map
Unstable (rustc_private)

analysis
Unstable (rustc_private)

tcx
Unstable (rustc_private)

trans
Unstable (rustc_private)