Trait rustc_driver::pretty::HirPrinterSupport

trait HirPrinterSupport<'hir>: PpAnn {
    fn sess(&self) -> &Session;
    fn hir_map(&self) -> Option<Map<'hir>>;
    fn pp_ann(&self) -> &dyn PpAnn;
}

Required Methods

fn sess(&self) -> &Session

Provides a uniform interface for re-extracting a reference to a Session from a value that now owns it.

fn hir_map(&self) -> Option<Map<'hir>>

Provides a uniform interface for re-extracting a reference to an hir_map::Map from a value that now owns it.

fn pp_ann(&self) -> &dyn PpAnn

Produces the pretty-print annotation object.

(Rust does not yet support upcasting from a trait object to an object for one of its supertraits.)