Crate rustc_driver

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The Rust compiler.

Note

This API is completely unstable and subject to change.

Modules

driver 	 [
Experimental
]

pretty 	 [
Experimental
]
The various pretty print routines.

profile 	 [
Experimental
]

target_features 	 [
Experimental
]

Structs

DefaultTransCrate 	 [
Experimental
]

RustcDefaultCalls 	 [
Experimental
]

Enums

Compilation 	 [
Experimental
]

Traits

CompilerCalls 	 [
Experimental
]

Functions

abort_on_err 	 [
Experimental
]

diagnostics_registry 	 [
Experimental
]

enable_save_analysis 	 [
Experimental
]

get_args 	 [
Experimental
]

handle_options 	 [
Experimental
]
Process command line options. Emits messages as appropriate. If compilation should continue, returns a getopts::Matches object parsed from args, otherwise returns None.

in_rustc_thread 	 [
Experimental
]
Runs f in a suitable thread for running rustc; returns a Result with either the return value of f or -- if a panic occurs -- the panic value.

main 	 [
Experimental
]

monitor 	 [
Experimental
]
Run a procedure which will detect panics in the compiler and print nicer error messages rather than just failing the test.

run 	 [
Experimental
]

run_compiler 	 [
Experimental
]

version 	 [
Experimental
]
Prints version information