Crate rustc_driver

Unstable (rustc_private)

The Rust compiler.

Note

This API is completely unstable and subject to change.

Modules

driver 	 [Unstable]

pretty 	 [Unstable]
The various pretty print routines.

target_features 	 [Unstable]

Structs

RustcDefaultCalls 	 [Unstable]

Enums

Compilation 	 [Unstable]

Traits

CompilerCalls 	 [Unstable]

Functions

abort_on_err 	 [Unstable]

commit_date_str 	 [Unstable]
Returns the "commit date" of HEAD of the Git repo from which rustc was built as a static string.

commit_hash_str 	 [Unstable]
Returns the full SHA1 hash of HEAD of the Git repo from which rustc was built.

diagnostics_registry 	 [Unstable]

handle_options 	 [Unstable]
Process command line options. Emits messages as appropriate. If compilation should continue, returns a getopts::Matches object parsed from args, otherwise returns None.

main 	 [Unstable]

monitor 	 [Unstable]
Run a procedure which will detect panics in the compiler and print nicer error messages rather than just failing the test.

release_str 	 [Unstable]
Returns a version string such as "0.12.0-dev".

run 	 [Unstable]

run_compiler 	 [Unstable]

run_compiler_with_file_loader 	 [Unstable]

version 	 [Unstable]
Prints version information