Struct rustc_driver::driver::CompileState

pub struct CompileState<'a, 'ast, 'tcx> {
    pub input: &'a Input,
    pub session: &'a Session,
    pub cfg: Option<&'a CrateConfig>,
    pub krate: Option<&'a Crate>,
    pub crate_name: Option<&'a str>,
    pub output_filenames: Option<&'a OutputFilenames>,
    pub out_dir: Option<&'a Path>,
    pub expanded_crate: Option<&'a Crate>,
    pub hir_crate: Option<&'a Crate>,
    pub ast_map: Option<&'a Map<'ast>>,
    pub mir_map: Option<&'a MirMap<'tcx>>,
    pub analysis: Option<&'a CrateAnalysis<'a>>,
    pub tcx: Option<&'a ctxt<'tcx>>,
    pub lcx: Option<&'a LoweringContext<'a>>,
    pub trans: Option<&'a CrateTranslation>,
}

Unstable (rustc_private)

State that is passed to a callback. What state is available depends on when during compilation the callback is made. See the various constructor methods (state_*) in the impl to see which data is provided for any given entry point.

Fields

input
Unstable (rustc_private)

session
Unstable (rustc_private)

cfg
Unstable (rustc_private)

krate
Unstable (rustc_private)

crate_name
Unstable (rustc_private)

output_filenames
Unstable (rustc_private)

out_dir
Unstable (rustc_private)

expanded_crate
Unstable (rustc_private)

hir_crate
Unstable (rustc_private)

ast_map
Unstable (rustc_private)

mir_map
Unstable (rustc_private)

analysis
Unstable (rustc_private)

tcx
Unstable (rustc_private)

lcx
Unstable (rustc_private)

trans
Unstable (rustc_private)